+++ packages/mint-components/src/components/sqm-navigation-sidebar/sqm-navigation-sidebar.tsx
@@ -1,4 +1,9 @@
+import {
+  useCurrentPage,
+  useProgramId,
+} from "@saasquatch/component-boilerplate";
 import { withHooks } from "@saasquatch/stencil-hooks";
+import { useEffect, useState } from "@saasquatch/universal-hooks";
 import { Component, h, State } from "@stencil/core";
 import { NavigationSidebarView } from "./sqm-navigation-sidebar-view";
 
@@ -20,10 +25,33 @@ export class NavigationSidebar {
   disconnectedCallback() {}
 
   render() {
+    const props = useNavigationSidebar();
+
     return (
-      <NavigationSidebarView>
+      <NavigationSidebarView {...props}>
         <slot />
       </NavigationSidebarView>
     );
   }
+}
+
+function useNavigationSidebar() {
+  const location = useCurrentPage();
+  const programId = useProgramId();
+
+  const [checked, setChecked] = useState(false);
+
+  // Close hamburger menu after navigation
+  useEffect(() => {
+    setChecked(false);
+  }, [location.pathname, programId]);
+
+  function onClick(e) {
+    setChecked(e.target.checked);
+  }
+
+  return {
+    checked,
+    onClick,
+  };
 }