+++ packages/mint-components/src/components/sqm-portal-frame/sqm-portal-frame-view.tsx
@@ -14,7 +14,6 @@ export interface PortalFrameViewProps {
 
 const style = {
   HostBlock: HostBlock,
-
   Frame: {
     display: "flex",
     "flex-direction": "column",
@@ -59,7 +58,10 @@ const styleString = sheet.toString();
 export function PortalFrameView(props: PortalFrameViewProps, children: VNode) {
   const { data, notFullScreen } = props;
   return (
-    <div style={{minHeight: notFullScreen ? "100%" : "100vh"}} class={sheet.classes.Frame}>
+    <div
+      style={{ minHeight: notFullScreen ? "100%" : "100vh" }}
+      class={sheet.classes.Frame}
+    >
       <style type="text/css">{styleString}</style>
       <div class={sheet.classes.HeaderWrapper}>
         {data.header}