+++ packages/mint-components/src/stories/NewPortal.stories.tsx
@@ -25,21 +25,6 @@ const style = {
 const sheet = createStyleSheet(style);
 const styleString = sheet.toString();
 
-const barProps = {
-  data: {
-    programs: [
-      {
-        key: "program1",
-        label: "My Referral Program",
-      },
-      {
-        key: "program2",
-        label: "My Rewards Program",
-      },
-    ],
-  },
-};
-
 const item1Props = {
   states: {
     active: false,
@@ -64,7 +49,7 @@ const item2Props = {
 
 const Sidebar = () => {
   return (
-    <NavigationSidebarView {...barProps}>
+    <NavigationSidebarView>
       <NavigationSidebarItemView {...item1Props} />
       <NavigationSidebarItemView {...item2Props} />
     </NavigationSidebarView>